// src/api/sqlmodule/model/validationCheckRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { ValidationCheckRequestKind } from './validationCheckRequestKind';

/**
 * Критерий в запросе сохранения конфигурации.
 */
export interface ValidationCheckRequest {
  /** @nullable */
  id?: string | null;
  kind?: ValidationCheckRequestKind;
  /** @nullable */
  value?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  order?: number | null;
}